services/requirements.txt
bottle
redis
gunicorn
jinja2
netifaces
pynacl
click